plot.compare_means {radiant.basics}	R Documentation

Plot method for the compare_means function

Description

Plot method for the compare_means function

Usage

## S3 method for class 'compare_means'
plot(x, plots = "scatter", shiny = FALSE,
  custom = FALSE, ...)

Arguments

x
Return value from compare_means

plots
One or more plots ("bar", "density", "box", or "scatter")

shiny
Did the function call originate inside a shiny app

custom
Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and http://docs.ggplot2.org for options.

...
further arguments passed to or from other methods

Details

See https://radiant-rstats.github.io/docs/basics/compare_means.html for an example in Radiant

See Also

compare_means to calculate results

summary.compare_means to summarize results

Examples

result <- compare_means(diamonds, "cut", "price")
plot(result, plots = c("bar", "density"))

[Package radiant.basics version 0.9.9 Index]