plot.single_prop {radiant.basics}	R Documentation

Plot method for the single_prop function

Description

Plot method for the single_prop function

Usage

## S3 method for class 'single_prop'
plot(x, plots = "bar", shiny = FALSE,
  custom = FALSE, ...)

Arguments

x
Return value from single_prop

plots
Plots to generate. "bar" shows a bar chart of the data. The "simulate" chart shows the location of the sample proportion and the comparison value (comp_value). Simulation is used to demonstrate the sampling variability in the data under the null-hypothesis

shiny
Did the function call originate inside a shiny app

custom
Logical (TRUE, FALSE) to indicate if ggplot object (or list of ggplot objects) should be returned. This option can be used to customize plots (e.g., add a title, change x and y labels, etc.). See examples and http://docs.ggplot2.org for options.

...
further arguments passed to or from other methods

Details

See https://radiant-rstats.github.io/docs/basics/single_prop.html for an example in Radiant

See Also

single_prop to generate the result

summary.single_prop to summarize the results

Examples

result <- single_prop(titanic, "survived", lev = "Yes", comp_value = 0.5,  alternative = "less")
plot(result, plots = c("bar", "simulate"))

[Package radiant.basics version 0.9.9 Index]