summary.goodness {radiant.basics}	R Documentation

Summary method for the goodness function

Description

Summary method for the goodness function

Usage

## S3 method for class 'goodness'
summary(object, check = "", dec = 2, ...)

Arguments

object
Return value from goodness

check
Show table(s) for the selected variable (var). "observed" for the observed frequencies table, "expected" for the expected frequencies table (i.e., frequencies that would be expected if the null hypothesis holds), "chi_sq" for the contribution to the overall chi-squared statistic for each cell (i.e., (o - e)^2 / e), "dev_std" for the standardized differences between the observed and expected frequencies (i.e., (o - e) / sqrt(e)), and "dev_perc" for the percentage difference between the observed and expected frequencies (i.e., (o - e) / e)

dec
Number of decimals to show

...
further arguments passed to or from other methods.

Details

See https://radiant-rstats.github.io/docs/basics/goodness for an example in Radiant

See Also

goodness to calculate results

plot.goodness to plot results

Examples

result <- goodness(newspaper, "Income", c(.3, .7))
summary(result, check = c("observed", "expected", "chi_sq"))
goodness(newspaper, "Income", "1/3 2/3") %>% summary("observed")

[Package radiant.basics version 0.9.9 Index]