correlation {radiant.basics}	R Documentation

Calculate correlations for two or more variables

Description

Calculate correlations for two or more variables

Usage

correlation(dataset, vars = "", method = "pearson", data_filter = "")

Arguments

dataset
Dataset

vars
Variables to include in the analysis. Default is all but character and factor variables with more than two unique values are removed

method
Type of correlations to calculate. Options are "pearson", "spearman", and "kendall". "pearson" is the default

data_filter
Expression entered in, e.g., Data > View to filter the dataset in Radiant. The expression should be a string (e.g., "price > 10000")

Details

See https://radiant-rstats.github.io/docs/basics/correlation.html for an example in Radiant

Value

A list with all variables defined in the function as an object of class compare_means

See Also

summary.correlation to summarize results

plot.correlation to plot results

Examples

correlation(diamonds, c("price", "carat")) %>% str()
correlation(diamonds, "x:z") %>% str()

[Package radiant.basics version 0.9.9 Index]